#189: Moved all apply_* funcs to 'apply.rs' mod
