+++ src/core/effect.rs
@@ -1,4 +1,4 @@
-use core::{ObjId, State, Strength};
+use core::Strength;
 
 // TODO: subturns? EffectTime?
 //
@@ -24,25 +24,3 @@ pub enum Effect {
 
 #[derive(Clone, Debug)]
 pub struct Wound(pub Strength);
-
-pub fn apply(state: &mut State, id: ObjId, effect: &Effect) {
-    debug!("effect::apply: {:?}", effect);
-    match *effect {
-        Effect::Kill => apply_kill(state, id),
-        Effect::Wound(ref effect) => apply_wound(state, id, effect),
-        Effect::Miss => apply_miss(state, id),
-    }
-}
-
-pub fn apply_kill(state: &mut State, id: ObjId) {
-    state.parts_mut().remove(id);
-}
-
-pub fn apply_wound(state: &mut State, id: ObjId, effect: &Wound) {
-    let damage = effect.0;
-    let strength = state.parts_mut().strength.get_mut(id);
-    strength.strength.0 -= damage.0;
-    assert!(strength.strength.0 > 0);
-}
-
-pub fn apply_miss(_: &mut State, _: ObjId) {}
